policy: Limit proxy_id to 16 bits
Proxy ID is really a port number, so we can safely limit it to 16 bits.

+++ cilium/websocket.cc
@@ -134,7 +134,7 @@ Network::FilterStatus Instance::onNewConnection() {
   std::string pod_ip;
   bool is_ingress;
   uint32_t identity, destination_identity;
-  uint32_t proxy_id;
+  uint16_t proxy_id;
 
   auto& conn = callbacks_->connection();
